Added application context

+++ servlets/src/main/java/ru/netology/controller/PostController.java
@@ -1,6 +1,7 @@
 package ru.netology.controller;
 
 import com.google.gson.Gson;
+import org.springframework.stereotype.Controller;
 import ru.netology.model.Post;
 import ru.netology.service.PostService;
 
@@ -8,6 +9,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.Reader;
 
+@Controller
 public class PostController {
     public static final String APPLICATION_JSON = "application/json";
     final Gson gson;
@@ -34,9 +36,11 @@ public class PostController {
     }
 
     public void removeById(long id, HttpServletResponse response) throws IOException {
-        service.getById(id);
-        service.removeById(id);
-        setContentTypeAndPrintReply(response, APPLICATION_JSON, "Post with id: " + id + " deleted successfully");
+        if (service.getById(id) != null) {
+            service.removeById(id);
+            setContentTypeAndPrintReply(response, APPLICATION_JSON, "Post with id: " + id + " deleted successfully");
+        }
+
 
     }
